refactor(order): Refactor order services: Move OrderCalculationService, OrderInventoryIntegrationService, OrderManagementService, and OrderService to new structure; introduce OrderValidationUtils for shared validation logic. Update imports and deprecate old service references.
