+++ lib/app/wiring/provider.dart
@@ -60,10 +60,10 @@ import "../../features/menu/services/menu_service.dart";
 import "../../features/order/models/order_model.dart" show Order, OrderItem;
 import "../../features/order/repositories/order_item_repository.dart";
 import "../../features/order/repositories/order_repository.dart";
-import "../../features/order/services/cart_management_service.dart";
-import "../../features/order/services/order_calculation_service.dart";
-import "../../features/order/services/order_inventory_integration_service.dart";
-import "../../features/order/services/order_management_service.dart";
+import "../../features/order/services/cart/cart_management_service.dart";
+import "../../features/order/services/order/order_calculation_service.dart";
+import "../../features/order/services/order/order_inventory_integration_service.dart";
+import "../../features/order/services/order/order_management_service.dart";
 import "../../infra/batch/batch_processing_service.dart" as batch_impl;
 import "../../infra/local/cache/memory_cache_adapter.dart";
 import "../../infra/local/cache/ttl_cache_adapter.dart";
