+++ lib/features/order/services/order_service.dart
@@ -1,3 +1,2 @@
 @Deprecated("OrderServiceはOrderManagementServiceへ統合されました。order_management_service.dartを参照してください。")
 export "order_management_service.dart";
-import "../../../core/constants/enums.dart";
